fix(telemetry): use int (not uint64) for TLS confidence counters
gosec G115 flagged three int→uint64 conversions in snapshotSummary on
the new TLS confidence counters. Every other counter in telemetry.Summary
is `int`; the new H8 fields were the only `uint64` outliers. Switch the
four fields to `int` for consistency with the rest of the struct and
remove the casts at the call site.

JSON wire format is unchanged (Go marshals both as numbers).

+++ internal/telemetry/telemetry.go
@@ -107,21 +107,21 @@ type Summary struct {
 	//   - unknown:  no usable SNI signal — includes the P4 KTLS override
 	//     (see TLSConfidenceUnknownKTLS in the digest input for the subset
 	//     attributed specifically to kernel-TLS offload).
-	TLSConfidenceFull             uint64 `json:"tls_confidence_full,omitempty"`
-	TLSConfidencePartial          uint64 `json:"tls_confidence_partial,omitempty"`
-	TLSConfidenceInferred         uint64 `json:"tls_confidence_inferred,omitempty"`
-	TLSConfidenceUnknown          uint64 `json:"tls_confidence_unknown,omitempty"`
-	ProcForkEvents                int    `json:"proc_fork_events,omitempty"`
-	Connect4TupleUpdateFailures   int    `json:"connect4_tuple_update_failures,omitempty"`
-	UDPRingbufReserveFailures     int    `json:"udp_ringbuf_reserve_failures,omitempty"`
-	DNSRingbufReserveFailures     int    `json:"dns_ringbuf_reserve_failures,omitempty"`
-	ConnectRingbufReserveFailures int    `json:"connect_ringbuf_reserve_failures,omitempty"`
-	HTTPRingbufReserveFailures    int    `json:"http_ringbuf_reserve_failures,omitempty"`
-	TLSRingbufReserveFailures     int    `json:"tls_ringbuf_reserve_failures,omitempty"`
-	ExecRingbufReserveFailures    int    `json:"exec_ringbuf_reserve_failures,omitempty"`
-	ForkRingbufReserveFailures    int    `json:"fork_ringbuf_reserve_failures,omitempty"`
-	FSRingbufReserveFailures      int    `json:"fs_ringbuf_reserve_failures,omitempty"`
-	UDPSendmsgMultiIovecObserved  int    `json:"udp_sendmsg_multi_iovec_observed,omitempty"`
+	TLSConfidenceFull             int `json:"tls_confidence_full,omitempty"`
+	TLSConfidencePartial          int `json:"tls_confidence_partial,omitempty"`
+	TLSConfidenceInferred         int `json:"tls_confidence_inferred,omitempty"`
+	TLSConfidenceUnknown          int `json:"tls_confidence_unknown,omitempty"`
+	ProcForkEvents                int `json:"proc_fork_events,omitempty"`
+	Connect4TupleUpdateFailures   int `json:"connect4_tuple_update_failures,omitempty"`
+	UDPRingbufReserveFailures     int `json:"udp_ringbuf_reserve_failures,omitempty"`
+	DNSRingbufReserveFailures     int `json:"dns_ringbuf_reserve_failures,omitempty"`
+	ConnectRingbufReserveFailures int `json:"connect_ringbuf_reserve_failures,omitempty"`
+	HTTPRingbufReserveFailures    int `json:"http_ringbuf_reserve_failures,omitempty"`
+	TLSRingbufReserveFailures     int `json:"tls_ringbuf_reserve_failures,omitempty"`
+	ExecRingbufReserveFailures    int `json:"exec_ringbuf_reserve_failures,omitempty"`
+	ForkRingbufReserveFailures    int `json:"fork_ringbuf_reserve_failures,omitempty"`
+	FSRingbufReserveFailures      int `json:"fs_ringbuf_reserve_failures,omitempty"`
+	UDPSendmsgMultiIovecObserved  int `json:"udp_sendmsg_multi_iovec_observed,omitempty"`
 	// SendmmsgMultiMessage counts NR_SENDMMSG calls with vlen>1 (mmsghdr vector
 	// length, distinct from per-message msg_iovlen). Messages 2..N are not
 	// introspected — non-zero quantifies the multi-message silent gap (BG-03).
